Changes 2024 part 6 - Following changes has been done ,
The PELL has a minimum limit of $740 that is not built into the current logic (apparently).  At this point, it’s probably best to just have Pell show as Zero if the SAI is greater than 6655 (that’s max-Pell $7395 minus SAI 6655 = Min-Pell $740).

The La Sierra Need Grant (LSNEED) of $1,000 needs to be added if the SAI is below 12,000 (right now it packages if the SAI is between 0 and 12,000, but it needs to also be included for negative SAI numbers).

The La Sierra Achievement Award (MERIT) – which is the award associated with SAT or ACT test scores – should be increased from $10,200 to $10,700.